chore(tags): extracted tags to own component
